+++ backend/src/main/java/com/project/backend/controller/BackendController.java
@@ -1,4 +1,4 @@
-package com.project.backend;
+package com.project.backend.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
