+++ backend/src/main/java/com/project/backend/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.project.backend;
+package com.project.backend.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
